pyproject.toml
[project]
name = "pair-tone"
version = "0.1.0"
description = "Annotation app for Japanese mishearing stimulus pairs"
readme = "annotation_app_README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.35",
    "boto3>=1.34",
    "pydantic>=2.7",
    "pandas>=2.2",
    "python-dotenv>=1.0",
    "orjson>=3.10",
    "tenacity>=8.3",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-mock>=3.14",
    "mypy>=1.10",
    "ruff>=0.5",
    "boto3-stubs[s3,polly]>=1.34",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
